Remove lantern light: it makes other sprites disappear
Since the lantern light texture is very big, other sprites will not be
rendered. After only 4 lanterns, the problems presents itself.
So, instead of using sprites for lantern light, tiles will have to be used.
Since the player's light should be of the highest quality and it's only
one, it can still be implemented using a sprite.