Add a respawn feature
Previously, the player would lose all progress when dying and would be
brought to the start menu. Now, the player can respawn in the same
world, losing all the items in the inventory. This means that any change
to the world (e.g. chests) will be retained after death.

TODO: a setting that lets the player keep his inventory after death.