Replace pseudo-random number generator with the one in 'libsimplegba'
The 'rand' and 'srand' functions have now been deleted and replaced by
'random' and 'random_seed'.

'random' asks for a bound, so that the returned value is restricted to
the given bound. This is practically equivalent to 'rand() % x', except
that 'random' does not use the modulo operator.

Performance is significantly improved in areas that use RNG (e.g. level
generation, tile ticking).