Storage: rewrite the entire system and change the save file layout
The new layout has these main differences from the previous one:
- Header and footer are now merged into a single 0.5 KB header
- Level arrays are now stored contiguously with arrays of the same type
  (e.g. all entities from all levels, then all tile data...)
- Tile IDs are stored using Run-Length Encoding (RLE)

Run-Length Encoding is much more efficient than a fixed-length encoding
when storing levels, since tiles of the same type tend to be clustered.
Using this encoding allows for up to 255 types of tile to exist without
any issue (e.g. for expansions or mods).

It is possible to run out of memory for tile IDs, since RLE is only
efficient if consecutive runs of the same tile are present. I believe it
is unlikely to happen, but dedicated players may attempt to break the
save system on purpose (e.g. by building a checkered pattern). No safety
checks are in place to prevent saving issues.